pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-env"
version = "1.0.0"
description = "Customer Support OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["support_env", "support_env.server", "server"]



